+++ src/components/SignUp.vue
@@ -9,7 +9,6 @@
 <script lang="ts">
 import { Component, Vue } from "vue-property-decorator";
 import { Getter, Action, namespace } from "vuex-class";
-import { IAuthorizationResponse, STORAGE_KEY_AUTH_STATE } from "@/domain/Qiita";
 
 const QiitaAction = namespace("QiitaModule", Action);
 const QiitaGetter = namespace("QiitaModule", Getter);
@@ -21,21 +20,5 @@ export default class SignUp extends Vue {
 
   @QiitaAction
   signUp!: () => void;
-
-  @QiitaAction
-  createAccount!: (query: object) => void;
-
-  created(): void {
-    const query: any = this.$route.query;
-    const params: IAuthorizationResponse = {
-      code: query.code,
-      callbackState: query.state,
-      localState:
-        window.localStorage.getItem(STORAGE_KEY_AUTH_STATE) || undefined
-    };
-
-    this.$router.push({ query: {} });
-    this.createAccount(params);
-  }
 }
 </script>
