+++ src/router.ts
@@ -4,6 +4,7 @@ import Counter from "./components/Counter.vue";
 import Weather from "./components/Weather.vue";
 import SignUp from "./components/SignUp.vue";
 import Login from "./components/Login.vue";
+import OAuthCallback from "./components/OAuthCallback.vue";
 import Error from "./components/Error.vue";
 import Home from "./views/Home.vue";
 
@@ -37,6 +38,11 @@ export default new Router({
       path: "/signup",
       name: "signup",
       component: SignUp
+    },
+    {
+      path: "/oauth/callback",
+      name: "oAuthCallback",
+      component: OAuthCallback
     },
     {
       path: "/error",
